Patch_COMPRESS_DATA

+++ contracts/etf/ETFProxy.sol
@@ -32,7 +32,7 @@ contract ETFProxy is
     */
     struct SwapInfo {
         IERC20Upgradeable token;
-        uint256 price;
+        uint64 price;
         bytes data;
     }
 
@@ -146,7 +146,9 @@ contract ETFProxy is
 
         require(_planDetail.active, "ETFProxy: Plan is not active.");
 
+
         _requiredValidSwapInfo(_swaps, _planDetail.tokenPercentages);
+
 
         if (!_tokenIn.isETH()) {
             _transferTokenFromTo(
@@ -165,6 +167,7 @@ contract ETFProxy is
 
         IETFReceipt.TokenDetail[]
             memory _tokenDetails = new IETFReceipt.TokenDetail[](_swaps.length);
+
         for (uint256 i = 0; i < _swaps.length; i++) {
             uint256 _slicedAmountIn = (_amount *
                 _planDetail.tokenPercentages[i].percentage) / MAX_P;
@@ -190,7 +193,7 @@ contract ETFProxy is
             IETFReceipt.TokenDetail memory _tokenDetail = IETFReceipt
                 .TokenDetail({
                     token: address(_swaps[i].token),
-                    amount: _amountOut,
+                    amount: uint128(_amountOut),
                     price: _swaps[i].price
                 });
 
@@ -199,10 +202,9 @@ contract ETFProxy is
 
         uint256 _investId = _ETFReceipt.mint(
             _userAddress,
-            _planDetail.id,
+            uint16(_planDetail.id),
             _tokenDetails
         );
-
         emit Invested(_userAddress, _investId, _planDetail.id, msg.sender);
     }
 
@@ -262,7 +264,7 @@ contract ETFProxy is
 
         if (_percentage != MAX_P) {
             _ETFReceipt.burnAndMint(
-                _tokenId,
+                uint32(_tokenId),
                 msg.sender,
                 _invest.planId,
                 _remainsTokenDetails
@@ -300,7 +302,7 @@ contract ETFProxy is
         for (uint i = 0; i < _invest.tokenDetails.length; i++) {
             address _token = _invest.tokenDetails[i].token;
             uint256 _amount = _invest.tokenDetails[i].amount;
-            uint256 _price = _invest.tokenDetails[i].price;
+            uint64 _price = _invest.tokenDetails[i].price;
 
             uint256 _beforeAmount = IERC20Upgradeable(_token).uniBalanceOf(
                 address(this)
@@ -328,7 +330,7 @@ contract ETFProxy is
                 IETFReceipt.TokenDetail memory _tokenDetail = IETFReceipt
                     .TokenDetail({
                         token: _token,
-                        amount: _remainsAmount,
+                        amount: uint128(_remainsAmount),
                         price: _price
                     });
 
@@ -338,7 +340,7 @@ contract ETFProxy is
 
         if (_percentage != MAX_P) {
             _ETFReceipt.burnAndMint(
-                _tokenId,
+                uint32(_tokenId),
                 msg.sender,
                 _invest.planId,
                 _tokenDetails
@@ -632,8 +634,8 @@ contract ETFProxy is
                 IETFReceipt.TokenDetail memory _tokenDetail = IETFReceipt
                     .TokenDetail({
                         token: _invest.tokenDetails[i].token,
-                        amount: _remainsAmount,
-                        price: _invest.tokenDetails[i].price
+                        amount: uint128(_remainsAmount),
+                        price: uint16(_invest.tokenDetails[i].price)
                     });
 
                 _tokenDetails[i] = _tokenDetail;
